requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.0
flasgger==0.9.5
Flask==2.2.3
Flask-Autodoc==0.1.2
flask-marshmallow==0.14.0
flask-paginate==2022.1.8
flask-rest-paginate==1.1.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mistune==2.0.5
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandasql==0.7.3
pluggy==1.0.0
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.4
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
python-dotenv==1.0.1
